Scripts/permissions/requirements.txt
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
boto3==1.4.4
botocore==1.5.85
certifi==2017.4.17
chardet==3.0.4
configparser==3.5.0
decorator==4.1.2
docutils==0.13.1
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.1.1
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==7.0.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
multiprocessing==2.6.2.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.0.0
pandocfilters==1.4.2
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.11.1
s3transfer==0.1.10
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
slacker==0.9.50
tableauserverclient==0.4.1
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.2
